* OnixProduct can process items with repeat tags.
OnixProduct used to raise an Exception when it encountered a record with more than one of the
same tag. This would happen, for example, with the <title> tag on some records, which is legal in
ONIX. If it encounters one of these cases, it will now return a list of OnixProducts instead of
raising an Exception. I also fixed some long wrapping lines, and added a module variable called
quiet, that will suppress output to stderr if set to True.